refactor and feedback addressing round 2
