+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/practitioner/Dstu2PractitionerTransformerTest.java
@@ -81,8 +81,8 @@ public class Dstu2PractitionerTransformerTest {
 
   @Test
   public void practitioner() {
-    assertThat(tx(DatamartPractitionerSamples.Datamart.create().practitioner()).toFhir())
-        .isEqualTo(Dstu2PractitionerSamples.Dstu2.create().practitioner());
+    assertThat(tx(PractitionerSamples.Datamart.create().practitioner()).toFhir())
+        .isEqualTo(PractitionerSamples.Dstu2.create().practitioner());
   }
 
   Dstu2PractitionerTransformer tx(DatamartPractitioner dm) {
